requirements.txt
pinecone
pyyaml
python-dotenv
llama-index
llama-index-vector-stores-pinecone
llama-index-embeddings-huggingface
llama-index-llms-groq
streamlit
